feat: alteracao da main para chamar novos metodos
Alteracao da main e padronizacao dos metodos
